\name{simpleqq}
\alias{simpleqq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plots a qq 'like' plot for lm object}
\description{
  Simple qq 'like' plot for lm object. This is not a true qq-plot but easier interpretation.
}
\usage{
simpleqq( model, ... )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{A regression model generated from lm}
  \item{...}{Other parameters to pass to plot}
}
\details{
  This function constructs a simple qqp 'like' lot for linear models.
}
\value{ }
\references{}
\author{Brian Keller}
\seealso{\code{\link{qqplot}}}
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }
